Improve response time for openning a link
Add a SamplesManager which will perform server side requests and cache the responses (in memory).
Introduce a new Box class which will be used as a class wrapper for value types.
Once the audio sample is available perform a server side call and cache the response to be used later.
When the user 'votes' use the SamplesManager to retrieve the cached response if available
otherwise do a network call.